bulk archive proposals
